update 12.19 2

+++ legged_gym/envs/base/legged_robot.py
@@ -102,7 +102,7 @@ class LeggedRobot(BaseTask):
         self.relative_pos[:] = self.root_states[:self.num_envs, 0:3] - self.env_origins[:]
         self.x_error = torch.square(self.relative_pos[:, 0] - self.cfg.env.desired_x)
         self.y_error = torch.square(self.relative_pos[:,1]-self.cfg.env.desired_y)
-        self.success = (self.x_error+self.y_error < 0.1).float()
+        self.success = (self.x_error+self.y_error < 0.1)
 
         self._post_physics_step_callback()
